Add algorithm to playGame-function
The playGame-function now takes both choice func-
tions and uses their return values to declare
the choice variables.

These are then checked to
match the correct available choices and if they
dont match, the function prints an message noti-
fying about this.

If the choices match, they are
then comparet to each other. If win conditions
for the player are met, a win-message is logged
and score incremented. If win-conditions are not
met the algorithm checks for losing conditions and
if these are met, logs the corresponding message
and increments the computers score.

If neither of these conditions are met, a tie is
assumed and a mesasge indicating this is logged.

Also minor changes, removed the extra calls for functions used
to debug, and the "hello world" message from the
beginning of the file.