Edit HTML and JS files to match lesson
I added the buttons and a div to keep the score in
the html. Then I made modified versions of the
functions in the JS file to suite what was reques-
ted in the lesson. After that i deleted the old
versions of functions. The buttons now play a
round of the game and the div displays the score