Add animatio, correct grammar, clean up code
Added a function that animates winning hand with
a tiny bounce. I also tweaked the color scheme.

Broke up a lot of the code into more functions.
Still a lot of trim left, ie. making the code
itself cleaner, simpler.

Made a function that makes sure the players choice
is capitalized in the log div.